# specs/openapi/ai-service.yaml
openapi: 3.1.0
info:
  title: Poetry RAG — AI Service API
  version: "0.4.0"
  description: |
    FastAPI AI 微服务对外暴露的接口。
    内部集成 LangGraph ReAct Agent、可配置 LLM（GLM / 豆包 / OpenAI / Ollama 等）、ChromaDB 向量检索、图像生成（CogView-4 等）。
    提供健康检查、异步生成、同步冒烟测试、思考流 SSE、LangGraph 对话 SSE、分镜多图 SSE 六类接口。
    LLM 和图像模型通过环境变量 `LLM_PROVIDER` / `IMAGE_PROVIDER` 自由配置，无需修改代码。
servers:
  - url: http://localhost:8000
    description: 本地开发
...
paths:
...
  /ai/api/v1/generate/async:
    post:
      summary: 异步启动 RAG 生成管道
      operationId: queueGeneration
      tags: [generate]
      description: |
        接收任务信息后立即返回 202，
        在后台执行 RAG 检索 → Prompt 增强 → 图像生成，
        完成后以 `POST callbackUrl` 回调 Backend。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerationRequest'
            example:
              taskId: "abc123-uuid"
              sourceText: "床前明月光，疑是地上霜"
              callbackUrl: "http://backend:8080/api/v1/poetry/callback"
              callbackToken: "poetry-callback-token-change-me"
      responses:
        "202":
          description: 任务已接受，后台处理中
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Task accepted"
                  taskId:
                    type: string

  # ── 同步简单生成（冒烟测试）───────────────────────────────────────────────
...
  /ai/api/v1/generate/think-stream:
    post:
      summary: GLM 推理过程 SSE 流
      operationId: thinkStream
      tags: [generate]
      description: |
        逐 token 推送 GLM 的 `reasoning_content`。
        适合前端实时展示"思考过程"。
        每个 SSE 事件格式：`data: {"text":"..."}`
        结束标记：`data: [DONE]`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimpleGenerationRequest'
      responses:
        "200":
          description: Server-Sent Events 流
          content:
            text/event-stream:
              schema:
                type: string

  # ── 对话 Session 管理 ──────────────────────────────────────────────────────
...
components:
  schemas:
    GenerationRequest:
      type: object
      required: [taskId, sourceText, callbackUrl, callbackToken]
      properties:
        taskId:
          type: string
          description: Backend 分配的任务 ID
        sourceText:
          type: string
          description: 原始诗句文本
        callbackUrl:
          type: string
          format: uri
          description: 完成后 POST 回调的地址
        callbackToken:
          type: string
          description: 回调鉴权 token，放入 X-Callback-Token 头
    SimpleGenerationRequest:
      type: object
      required: [sourceText]
      properties:
        sourceText:
          type: string
          description: 原始诗句文本
...